requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Pillow==9.5.0
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.19.5
requests>=2.27.1